fix(HomeSidePanel): 修复音乐/日历卡片编译错误
- java.time plusDays 需 Long 参数(Int 变量显式 toLong)
- object 单例内不允许 companion object(改为对象私有常量)
- collectAsState 委托属性不能 smart cast(转为局部变量)

+++ app/src/main/java/dev/ujhhgtg/wekit/features/items/beautify/home_screen_panel/HomeSidePanelCalendar.kt
@@ -286,7 +286,7 @@ private fun MonthGrid(
                         contentAlignment = Alignment.Center,
                     ) {
                         if (index >= leadingBlanks && index < leadingBlanks + daysInMonth) {
-                            val date = firstDay.plusDays(index - leadingBlanks)
+                            val date = firstDay.plusDays((index - leadingBlanks).toLong())
                             DayCell(
                                 day = date.dayOfMonth,
                                 selected = date == selectedDate,
@@ -310,7 +310,7 @@ private fun WeekRow(
     Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
         Row(modifier = Modifier.fillMaxWidth()) {
             repeat(7) { offset ->
-                val date = weekStart.plusDays(offset)
+                val date = weekStart.plusDays(offset.toLong())
                 val day = date.dayOfWeek.getDisplayName(
                     java.time.format.TextStyle.NARROW,
                     java.util.Locale.getDefault(),
